+++ app/components/PreferenceOption/tests/index.test.js
@@ -5,13 +5,17 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { PreferenceOption } from '../PreferenceOption';
+import { PreferenceOption } from '..';
 
 describe('<PreferenceOption />', () => {
   const mockSetPreferenceSelected = jest.fn();
   const rendered = shallow(
     <PreferenceOption
       preference="metrics"
+      label={{
+        id: 'app.containers.AgentPreferencesMenu.metrics',
+        defaultMessage: 'Metrics',
+      }}
       setPreferenceSelected={mockSetPreferenceSelected}
     />
   );