+++ app/containers/AgentPreferencesMenu/index.js
@@ -7,13 +7,18 @@ import { injectIntl } from 'react-intl';
 import PopupDialog from 'components/PopupDialog';
 import AgentStatsMenu from 'containers/AgentStatsMenu';
 import AgentNotificationsMenu from 'containers/AgentNotificationsMenu';
+import AudioOutputMenu from 'containers/AudioOutputMenu';
 import AgentTransferMenuPreferenceMenu from 'containers/AgentTransferMenuPreferenceMenu';
 import ErrorBoundary from 'components/ErrorBoundary';
 
+import PreferenceOption from 'components/PreferenceOption';
+import { isAlpha } from 'utils/url';
+import { selectOutputSelectionSupported } from 'containers/AudioOutputMenu/selectors';
 import { selectHasViewStatsPermission } from './selectors';
 
 import PreferenceTitle from './PreferenceTitle';
-import PreferenceOption from './PreferenceOption';
+
+import messages from './messages';
 
 const styles = {
   menu: {
@@ -38,6 +43,14 @@ export class AgentPreferencesMenu extends React.Component {
     this.setState({ preferenceSelected });
   };
 
+  renderPreferenceOption = (preference) => (
+    <PreferenceOption
+      preference={preference}
+      label={messages[preference]}
+      setPreferenceSelected={this.setPreferenceSelected}
+    />
+  );
+
   render() {
     let content;
     switch (this.state.preferenceSelected) {
@@ -49,29 +62,30 @@ export class AgentPreferencesMenu extends React.Component {
         content = <AgentNotificationsMenu />;
         break;
       }
+      case 'audioOutput': {
+        content = <AudioOutputMenu />;
+        break;
+      }
       case 'transferMenu': {
         content = <AgentTransferMenuPreferenceMenu />;
         break;
       }
-      default:
+      default: {
         content = (
           <Fragment>
-            {this.props.hasViewStatsPermission && (
-              <PreferenceOption
-                preference="metrics"
-                setPreferenceSelected={this.setPreferenceSelected}
-              />
-            )}
-            <PreferenceOption
-              preference="notifications"
-              setPreferenceSelected={this.setPreferenceSelected}
-            />
-            <PreferenceOption
-              preference="transferMenu"
-              setPreferenceSelected={this.setPreferenceSelected}
-            />
+            {this.props.hasViewStatsPermission &&
+              this.renderPreferenceOption('metrics')}
+
+            {this.renderPreferenceOption('notifications')}
+
+            {isAlpha() &&
+              this.props.isOutputSelectionSupported &&
+              this.renderPreferenceOption('audioOutput')}
+
+            {this.renderPreferenceOption('transferMenu')}
           </Fragment>
         );
+      }
     }
 
     return (
@@ -97,10 +111,12 @@ AgentPreferencesMenu.propTypes = {
   isVisible: PropTypes.bool,
   hideMenu: PropTypes.func.isRequired,
   hasViewStatsPermission: PropTypes.bool,
+  isOutputSelectionSupported: PropTypes.bool.isRequired,
 };
 
 const mapStateToProps = (state, props) => ({
   hasViewStatsPermission: selectHasViewStatsPermission(state, props),
+  isOutputSelectionSupported: selectOutputSelectionSupported(state, props),
 });
 
 export default ErrorBoundary(