+++ app/containers/AudioOutputMenu/messages.js
@@ -1,0 +1,24 @@
+/*
+ * Copyright © 2015-2020 Serenova, LLC. All rights reserved.
+ */
+
+/*
+ * AudioOutputMenu Messages
+ *
+ */
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  media: {
+    id: 'app.containers.AudioOutputMenu.media',
+    defaultMessage: 'Media',
+  },
+  notifications: {
+    id: 'app.containers.AudioOutputMenu.notifications',
+    defaultMessage: 'Notifications',
+  },
+  voice: {
+    id: 'app.containers.AudioOutputMenu.voice',
+    defaultMessage: 'Voice',
+  },
+});